refactor(web): give positions their own file, and their ids a type
Two things a reviewer would have had to take on trust.

`lib/app/hooks.ts` imported `useChatSessions` for the document title, and
`useChatSessions` needed the position, so the two imported each other. It
resolved, because both are only read inside function bodies, but it would
have stopped resolving the moment either was read at module scope, and the
failure would have been an undefined at import time. The position moves to
`lib/app/position.ts`, which imports neither.

Ids came back as text because that is how they arrive in the URL, so every
caller that wanted a row parsed one: `Number.parseInt` here, `String(id) ===`
there. The parse happens once, in the derivation, and the type says which
ids are rows and which are opaque — chat ids stay strings.

That also removes the check `AgentViewer` was making by hand. Text that is
not a row id names nothing, and now reads the same as the parameter being
absent, so there is no separate "is this even a number" question to ask.

Both jest mocks follow the move. Left behind they would have silently
stopped mocking anything.
