+++ web/src/app/nrf/NRFChrome.tsx
@@ -9,7 +9,8 @@ import { Button, LineItemButton, OpenButton, Popover } from "@opal/components";
 import { SvgBubbleText, SvgSearchMenu, SvgSidebar } from "@opal/icons";
 import MinimalMarkdown from "@/components/chat/MinimalMarkdown";
 import { useIsSearchModeAvailable } from "@/lib/settings/hooks";
-import { useAppPosition, useCustomFooterContent } from "@/lib/app/hooks";
+import { useCustomFooterContent } from "@/lib/app/hooks";
+import { useAppPosition } from "@/lib/app/position";
 import type { AppMode } from "@/providers/QueryControllerProvider";
 import { useQueryController } from "@/providers/QueryControllerProvider";
 import { useTierAtLeast } from "@/hooks/useTierAtLeast";
