+++ web/src/layouts/chromes/AppChrome.tsx
@@ -64,11 +64,8 @@ import {
 } from "@opal/icons";
 import { useIsSearchModeAvailable, useSettings } from "@/lib/settings/hooks";
 import type { AppMode } from "@/providers/QueryControllerProvider";
-import {
-  useAppDocumentTitle,
-  useAppPosition,
-  useCustomFooterContent,
-} from "@/lib/app/hooks";
+import { useAppDocumentTitle, useCustomFooterContent } from "@/lib/app/hooks";
+import { useAppPosition } from "@/lib/app/position";
 import { useQueryController } from "@/providers/QueryControllerProvider";
 import { useTierAtLeast } from "@/hooks/useTierAtLeast";
 import { Tier } from "@/lib/settings/types";
