+++ web/src/sections/sidebar/ChatButton.tsx
@@ -29,7 +29,7 @@ import {
 } from "@/lib/sidebar/utils";
 import { handleMoveOperation } from "@/lib/sidebar/svc";
 import ButtonRenaming from "@/refresh-components/buttons/ButtonRenaming";
-import { useAppPosition } from "@/lib/app/hooks";
+import { useAppPosition } from "@/lib/app/position";
 import {
   SvgChevronLeft,
   SvgEdit,
